✨ Feat : 백그라운드에서 sse 연결 유지
